requirements.txt
numpy>=1.21.0
scikit-learn>=1.0.0
pygame>=2.0.0
matplotlib>=3.5.0
pandas>=1.3.0
pillow>=10.0.0
opencv-python>=4.8.0
